requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
chromadb==0.4.24
pydantic==2.8.0
pydantic-settings==2.4.0
openai==1.40.0
google-cloud-storage==2.18.0
pdfplumber==0.11.0
python-multipart==0.0.9
httpx==0.27.0
python-dotenv==1.0.0
hnswlib==0.7.0
numpy==1.26.4

